pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "connectchain"
version = "0.0.1"
description = "Enterprise-focused Generative AI Toolset built upon Langchain"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "American Express"},
]
keywords = ["langchain", "connectchain", "enterprise"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.9.2",
    "openai>=1.0.0",
    "pyyaml==6.0.1",
    "SQLAlchemy==2.0.22",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.24",
    "langchain-community>=0.3.26",
    "tiktoken>=0.7.0",
    "python-dotenv~=1.0.0",
    "pyopenssl==23.3.0",
    "langchain-mcp-adapters>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/americanexpress/connectchain"
Repository = "https://github.com/americanexpress/connectchain"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pylint>=3.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
]

[tool.pytest.ini_options]
testpaths = ["connectchain/test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
pythonpath = ["."]

[tool.coverage.run]
source = ["connectchain"]
omit = [
    "connectchain/test/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["connectchain"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
# Exclude test files from strict type checking
exclude = [
    "connectchain/test/.*",
    "connectchain/examples/.*",
]

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "openai.*",
    "tiktoken.*",
    "OpenSSL.*",
    "langchain_mcp_adapters.*",
]
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["connectchain"] 